Create Chapter 3 Challenge 4
Write the pseudocode for a program where the player and the computer trade places in the number guessing game. That is, the player picks a random number 1 and 100 that the computer has to guess. Before you start, think about how you guess. If all goes well, try coding the game.